skb_drop: move event to events mod
