+++ src/events/mod.rs
@@ -25,3 +25,4 @@ pub(crate) mod ct;
 pub(crate) mod nft;
 pub(crate) mod ovs;
 pub(crate) mod skb;
+pub(crate) mod skb_drop;
