+++ src/module/skb_drop/mod.rs
@@ -7,9 +7,7 @@
 pub(crate) mod skb_drop;
 pub(crate) use skb_drop::*;
 
-pub(crate) mod event;
-pub(crate) use event::*;
-
+mod bpf;
 mod skb_drop_hook {
     include!("bpf/.out/skb_drop_hook.rs");
 }
